Create README.md
Sequence style game for android devices
Sequence is a board game played with two decks of cards, including jokers, and 3 different colored coin tokens.  

The board consists of a grid 10 x 10 filled with the cards from two standard decks. - But thats only 90! two decks in 104 or 108
with jokers.  Well wild spaces in the 4 corners (96) and Jacks and Jokers are not on the board.  (104 - 8 = 96, 108 - 12 = 96)
Jacks instead act as jokers albeit with a twist: One-eyed jacks remove a game piece.
Two-eyed jacks and jokers allow a piece to be played anywhere.